Fix bug whereby NaNs propagate through labels when no events are found at all
In the `_filter_area` method, there is a line that adds the maximum label in
the previous timestep to the labels of the current timestep, so that all
of the labels become unique. In the edge case where there is no labelled
event present in the previous iteration, `labels[i-1,:,:]`, however, this
operation `labels[i-1,:,;].max().values` would return `np.nan`, which then
corrupts valid (integer) label values in the current iteration. This then
propagates through to the remainder of the dataset.

This bug is fixed by simply adding a line that sets the `max_label` to `0`
if it is a `np.nan` type.

Presumably this had been missed because ocetrac was only tested on global datasets, where at least one event is present at any given time. This is not the case the problem that @marianatorres4 is working on, where events dissappear from
the entire domain in winter.