Robust fix to NaN labels, regardless of when no-event iterations occur
In the previous commits of this PR, I had hard-coded the fix to the NaN max labels
to be a label value of 0.0, but this is only correct if the quiescent (no-event)
iteration happens at the beginning of the tracking period. If no-event iterations
occur in the middle of the tracking period, the max_label should instead pick up
from where it was last. I have refactored this bug-fix so that we keep track
of the current maximum label and replace NaN values with that.

This gives slightly different results than the previous bug fix and should be more
robust!